changed assignment statements

+++ cmpiler_lab_5/test.c
@@ -3,7 +3,7 @@ int foo(int a,int b){
 }
 int main(int ami){
 
-	float e;
+	int e;
 	e = .25;
 	int i;
 	for(e;e<10;++e){
